# Conflicts:
#	gieFrontend/src/main/java/org/giefront/Controller/EditDeleteCell.java
#	gieFrontend/src/main/java/org/giefront/Controller/EditDeleteCellEnt.java
#	gieFrontend/src/main/java/org/giefront/Controller/EntropriseModificationController.java
#	gieFrontend/src/main/java/org/giefront/Controller/FournisseurPersoControlleur.java
#	gieFrontend/src/main/java/org/giefront/Controller/IEntropriseModificationController.java
#	gieFrontend/src/main/java/org/giefront/FournisseurPersoController.java
#	gieFrontend/src/main/resources/org/Interfaces/EntrepriseModification.fxml
#	gieFrontend/src/main/resources/org/Interfaces/PersonneModification.fxml
#	gieFrontend/src/main/resources/org/Interfaces/entreprise.fxml
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurEntro.fxml
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurPerso.fxml

+++ gieFrontend/src/main/java/org/giefront/Controller/AdvancedSearchP.java
@@ -119,7 +119,7 @@ public class AdvancedSearchP implements Initializable {
     public void onDelete(ActionEvent event) throws IOException {
         Personne selectedPerson =  tableView_P.getSelectionModel().getSelectedItem();
         if (selectedPerson != null) {
-            ps.deletePersonne(Math.toIntExact(selectedPerson.getId()));
+            ps.deletePersonne((long) Math.toIntExact(selectedPerson.getId()));
             tableView_P.getItems().remove(selectedPerson);
             showAlert("Suppression reussie");
         }
